WebChartAxis

One of a chart's axes. set_axis() builds these for you. Every axis must carry type = "chartAxis", and the count has to match the chart type's - two for most, one for a gauge, none for a pie.

Usage

WebChartAxis(
  type = NA_character_,
  visible = NA,
  isLogarithmic = NA,
  title = NULL,
  valueFormat = NumberFormatOptions(),
  minimum = NA_real_,
  maximum = NA_real_,
  grid = NULL,
  guides = list(),
  lineSymbol = NULL,
  labels = WebChartText(),
  scrollbar = NULL,
  displayZeroLine = NA,
  integerOnlyValues = NA,
  displayCursorTooltip = NA,
  buffer = NA,
  tickSpacing = NA_real_,
  dateBaseInterval = NULL
)

Arguments

type

String.

visible

Bool.

isLogarithmic

Bool.

title

NULL or a WebChartText object.

valueFormat

A NumberFormatOptions object or a DateTimeFormatOptions object or a CategoryFormatOptions object.

minimum

Number.

maximum

Number.

grid

NULL or a ISimpleLineSymbol object.

guides

List.

lineSymbol

NULL or a ISimpleLineSymbol object.

labels

A WebChartText object.

scrollbar

NULL or a WebChartAxisScrollBar object.

displayZeroLine

Bool.

integerOnlyValues

Bool.

displayCursorTooltip

Bool.

buffer

Bool.

tickSpacing

Number.

dateBaseInterval

NULL or a TimeIntervalInfo object.

Value

An object of class WebChartAxis.

Examples

WebChartAxis(
  type = "chartAxis",
  minimum = 0,
  maximum = 5000,
  integerOnlyValues = TRUE,
  title = WebChartText(
    type = "chartText",
    visible = TRUE,
    content = WebChartTextSymbol(type = "esriTS", text = "Body mass (g)")
  )
)
#> <arcgisviz::WebChartAxis>
#>  @ type                : chr "chartAxis"
#>  @ visible             : logi NA
#>  @ isLogarithmic       : logi NA
#>  @ title               : <arcgisviz::WebChartText>
#>  .. @ type   : chr "chartText"
#>  .. @ visible: logi TRUE
#>  .. @ content: <arcgisviz::WebChartTextSymbol>
#>  .. .. @ type               : chr "esriTS"
#>  .. .. @ style              : chr NA
#>  .. .. @ text               : chr "Body mass (g)"
#>  .. .. @ color              : NULL
#>  .. .. @ backgroundColor    : NULL
#>  .. .. @ borderLineSize     : num NA
#>  .. .. @ borderLineColor    : NULL
#>  .. .. @ haloSize           : num NA
#>  .. .. @ haloColor          : NULL
#>  .. .. @ verticalAlignment  : <arcgisviz::WebChartTextSymbolVerticalAlignment>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:4] "baseline" "bottom" "middle" "top"
#>  .. .. .. @ allow_na: logi TRUE
#>  .. .. @ horizontalAlignment: <arcgisviz::WebChartTextSymbolHorizontalAlignment>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:4] "center" "justify" "left" "right"
#>  .. .. .. @ allow_na: logi TRUE
#>  .. .. @ rightToLeft        : logi NA
#>  .. .. @ kerning            : logi NA
#>  .. .. @ font               : NULL
#>  .. .. @ angle              : num NA
#>  .. .. @ xoffset            : chr NA
#>  .. .. @ yoffset            : chr NA
#>  @ valueFormat         : <arcgisviz::NumberFormatOptions>
#>  .. @ type       : chr NA
#>  .. @ intlOptions: <arcgisviz::IntlNumberFormatOptions>
#>  .. .. @ localeMatcher           : <arcgisviz::IntlLocaleMatcher>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:2] "best fit" "lookup"
#>  .. .. .. @ allow_na: logi TRUE
#>  .. .. @ style                   : <arcgisviz::IntlNumberFormatOptionsStyle>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:4] "currency" "decimal" "percent" "unit"
#>  .. .. .. @ allow_na: logi TRUE
#>  .. .. @ currency                : chr NA
#>  .. .. @ currencyDisplay         : <arcgisviz::IntlNumberFormatOptionsCurrencyDisplay>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:4] "code" "name" "narrowSymbol" "symbol"
#>  .. .. .. @ allow_na: logi TRUE
#>  .. .. @ useGrouping             : logi NA
#>  .. .. @ minimumIntegerDigits    : num NA
#>  .. .. @ minimumFractionDigits   : num NA
#>  .. .. @ maximumFractionDigits   : num NA
#>  .. .. @ minimumSignificantDigits: num NA
#>  .. .. @ maximumSignificantDigits: num NA
#>  .. .. @ numberingSystem         : chr NA
#>  .. .. @ compactDisplay          : <arcgisviz::IntlNumberFormatOptionsCompactDisplay>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:2] "long" "short"
#>  .. .. .. @ allow_na: logi TRUE
#>  .. .. @ notation                : <arcgisviz::IntlNumberFormatOptionsNotation>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:4] "compact" "engineering" "scientific" ...
#>  .. .. .. @ allow_na: logi TRUE
#>  .. .. @ signDisplay             : <arcgisviz::IntlNumberFormatOptionsSignDisplay>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:4] "always" "auto" "exceptZero" "never"
#>  .. .. .. @ allow_na: logi TRUE
#>  .. .. @ unit                    : chr NA
#>  .. .. @ unitDisplay             : <arcgisviz::IntlFormatWidth>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:3] "long" "narrow" "short"
#>  .. .. .. @ allow_na: logi TRUE
#>  .. .. @ currencySign            : <arcgisviz::IntlNumberFormatOptionsCurrencySign>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:2] "accounting" "standard"
#>  .. .. .. @ allow_na: logi TRUE
#>  @ minimum             : num 0
#>  @ maximum             : num 5000
#>  @ grid                : NULL
#>  @ guides              : list()
#>  @ lineSymbol          : NULL
#>  @ labels              : <arcgisviz::WebChartText>
#>  .. @ type   : chr NA
#>  .. @ visible: logi NA
#>  .. @ content: <arcgisviz::WebChartTextSymbol>
#>  .. .. @ type               : chr NA
#>  .. .. @ style              : chr NA
#>  .. .. @ text               : chr NA
#>  .. .. @ color              : NULL
#>  .. .. @ backgroundColor    : NULL
#>  .. .. @ borderLineSize     : num NA
#>  .. .. @ borderLineColor    : NULL
#>  .. .. @ haloSize           : num NA
#>  .. .. @ haloColor          : NULL
#>  .. .. @ verticalAlignment  : <arcgisviz::WebChartTextSymbolVerticalAlignment>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:4] "baseline" "bottom" "middle" "top"
#>  .. .. .. @ allow_na: logi TRUE
#>  .. .. @ horizontalAlignment: <arcgisviz::WebChartTextSymbolHorizontalAlignment>
#>  .. .. .. @ value   : chr NA
#>  .. .. .. @ variants: chr [1:4] "center" "justify" "left" "right"
#>  .. .. .. @ allow_na: logi TRUE
#>  .. .. @ rightToLeft        : logi NA
#>  .. .. @ kerning            : logi NA
#>  .. .. @ font               : NULL
#>  .. .. @ angle              : num NA
#>  .. .. @ xoffset            : chr NA
#>  .. .. @ yoffset            : chr NA
#>  @ scrollbar           : NULL
#>  @ displayZeroLine     : logi NA
#>  @ integerOnlyValues   : logi TRUE
#>  @ displayCursorTooltip: logi NA
#>  @ buffer              : logi NA
#>  @ tickSpacing         : num NA
#>  @ dateBaseInterval    : NULL